+++ backend/proofchain/src/main/java/com/proofchain/plan/PlansRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PlansRepository extends JpaRepository<Plans, Integer> {
+public interface PlansRepository extends JpaRepository<Plans, Long> {
 
     Optional<Plans> findByName(String name);
 }
